+++ Source/Proxy/Proxy+Root.swift
@@ -8,7 +8,7 @@
 import Foundation
 
 extension Proxy {
-    static func rootProxy<T>(context: Context) -> Proxy<T> {
+    static func rootProxy<T: Codable>(context: Context) -> Proxy<T> {
         return Proxy<T>(context: context, objectId: .root, path: [], value: {
             let object = context.getObject(objectId: .root)
