+++ Source/Proxy/Proxy.swift
@@ -8,7 +8,7 @@
 import Foundation
 
 @dynamicMemberLookup
-public final class Proxy<Wrapped: Codable> {
+public final class Proxy<Wrapped> {
 
     init(
         context: Context,
@@ -121,13 +121,6 @@ public final class Proxy<Wrapped: Codable> {
             let path = Array(self.path.dropLast())
             context.setListIndex(path: path, index: index, value: newValue)
         }
-    }
-
-    public func set(_ newValue: Wrapped) {
-        let mapper = TypeToObject()
-        let object = try! mapper.map(newValue)
-
-        set(newValue: object)
     }
     
 }
